chore: cleanup typings
