+++ tsd_typings/es6-promise/es6-promise.d.ts
