--- typings/custom.system.d.ts
@@ -1,14 +1,0 @@
-// see https://github.com/systemjs/systemjs/issues/321
-
-interface System {
-  import(name: string): any;
-  defined: any;
-  amdDefine: () => void;
-  amdRequire: () => void;
-  baseURL: string;
-  paths: { [key: string]: string };
-  meta: { [key: string]: Object };
-  config: any;
-}
-
-declare var System: System;
